requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
pydantic>=2.9.2
pydantic-settings>=2.5.2
pandas>=3.0.0
openpyxl>=3.1.5
python-multipart>=0.0.12
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
